refactor: extract validation rules analysis hooks/helpers
